Fixed type error

+++ src/main/java/presentation/SearchController.java
@@ -102,7 +102,7 @@ public class SearchController implements Initializable {
     }
 
     @FXML
-    void commitIDchange(TableColumn.CellEditEvent<Production, Long> event) {
+    void commitIDchange(TableColumn.CellEditEvent<Production, Integer> event) {
         int row = event.getTablePosition().getRow();
         Production production = event.getTableView().getItems().get(row);
